+++ snapatac2-python/snapatac2/tools/embedding/__init__.py
@@ -1,3 +1,2 @@
 from ._spectral import spectral
-from ._laplacian import laplacian
 from ._umap import umap